Criando a arvore de diretórios e finalizando a refatoração da SplashScreenActivity.java
