--- app/src/main/java/com/fullscreen/SplashScreenActivity.java
@@ -1,50 +1,0 @@
-package com.fullscreen;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
-import android.view.MenuItem;
-
-import com.example.webview.R;
-
-import java.util.Timer;
-import java.util.TimerTask;
-
-public class SplashScreenActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash_screen);
-        if (getSupportActionBar() != null)
-            getSupportActionBar().hide();
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_splash_screen, menu);
-
-        new Timer().schedule(new TimerTask() {
-
-            @Override
-            public void run() {
-                finish();
-                Intent intent = new Intent();
-                intent.setClass(SplashScreenActivity.this, MainActivity.class);
-                startActivity(intent);
-            }
-        }, 4000);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_settings:
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-}
