+++ app/src/main/java/com/fullscreen/utils/Constants.java
@@ -1,0 +1,9 @@
+package com.fullscreen.utils;
+
+/**
+ * Created by jackson on 02/11/15.
+ */
+public class Constants {
+    public static final long SPLASH_SCREEN_DURATION = 4000;
+    public static final long SPLASH_SCREEN_INTERVAL = 1000;
+}
